Added the first draft of the AI.
Currently, there is a lot of AI debugging output; just redirect it to a file if it bugs you. I think most of the imbalances in gameplay now are more an issue with map gen rather than AI. Current game setup is you as player 1 versus an AI as player 2.